D0/ex02/requirements.txt
customtkinter
Pillow
piexif
argparse
